chore: minor renaming stuff
