+++ .github/workflows/validation_pr.yaml
@@ -15,8 +15,6 @@ jobs:
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         with:
-          # Configure which types are allowed (newline-delimited).
-          # Default: https://github.com/commitizen/conventional-commit-types
           types: |
             fix
             feat
@@ -29,12 +27,7 @@ jobs:
             ci
             chore
             revert
-          # Configure additional validation for the subject based on a regex.
-          # This example ensures the subject doesn't start with an uppercase character & has more than 50 chars.
           subjectPattern: ^(?![A-Z])(?=.{1,50}$).+$
-          # If `subjectPattern` is configured, you can use this property to override
-          # the default error message that is shown when the pattern doesn't match.
-          # The variables `subject` and `title` can be used within the message.
           subjectPatternError: |
             The subject "{subject}" found in the pull request title "{title}"
             didn't match the configured pattern. Please ensure that the subject
@@ -49,6 +42,4 @@ jobs:
       - name: List files with merge conflict markers
         run: git --no-pager grep "<<<<<<<" ":(exclude).github/" || true
       - name: Fail or succeed job if any files with merge conflict markers have been checked in
-        # Find lines containing "<<<<<<<", then count the number of lines.
-        # 0 matching lines results in exit code 0, i.e. success.
         run: exit $(git grep "<<<<<<<" ":(exclude).github/" | wc --lines)
