+++ tests/test_unit/test_service/assertThrows.py
