+++ tests/test_unit/test_service/test_filer_http_pytest.py
@@ -7,7 +7,7 @@ from requests import Response
 from tesk.service.filer import HTTPTransput, Type
 
 PATH_DOWN = 'test_download_file.txt'
-PATH_UP = 'tests/test_unit/service/test_filer_http_pytest.py'
+PATH_UP = 'tests/test_unit/test_service/test_filer_http_pytest.py'
 SUCCESS = 200
 FAIL = 300
 URL = 'http://www.foo.bar'
